+++ packages/client/src/coordinator/connection/client.ts
@@ -504,9 +504,6 @@ export class StreamClient {
         response,
       },
     );
-    this.logger('trace', `client:${type} - Response payload`, {
-      response,
-    });
   };
 
   _logApiError = (type: string, url: string, error: unknown) => {
@@ -562,9 +559,9 @@ export class StreamClient {
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
     } catch (e: any /**TODO: generalize error types  */) {
       e.client_request_id = requestConfig.headers?.['x-client-request-id'];
-      this._logApiError(type, url, e);
       this.consecutiveFailures += 1;
       if (e.response) {
+        this._logApiError(type, url, e.response);
         /** connection_fallback depends on this token expiration logic */
         if (
           e.response.data.code === KnownCodes.TOKEN_EXPIRED &&
@@ -578,6 +575,7 @@ export class StreamClient {
         }
         return this.handleResponse(e.response);
       } else {
+        this._logApiError(type, url, e);
         // eslint-disable-next-line no-throw-literal
         throw e as AxiosError<APIErrorResponse>;
       }
@@ -590,16 +588,28 @@ export class StreamClient {
     });
   };
 
-  put = <T, D = unknown>(url: string, data?: D) => {
-    return this.doAxiosRequest<T, D>('put', url, data);
+  put = <T, D = unknown>(
+    url: string,
+    data?: D,
+    params?: AxiosRequestConfig['params'],
+  ) => {
+    return this.doAxiosRequest<T, D>('put', url, data, { params });
   };
 
-  post = <T, D = unknown>(url: string, data?: D) => {
-    return this.doAxiosRequest<T, D>('post', url, data);
+  post = <T, D = unknown>(
+    url: string,
+    data?: D,
+    params?: AxiosRequestConfig['params'],
+  ) => {
+    return this.doAxiosRequest<T, D>('post', url, data, { params });
   };
 
-  patch = <T, D = unknown>(url: string, data?: D) => {
-    return this.doAxiosRequest<T, D>('patch', url, data);
+  patch = <T, D = unknown>(
+    url: string,
+    data?: D,
+    params?: AxiosRequestConfig['params'],
+  ) => {
+    return this.doAxiosRequest<T, D>('patch', url, data, { params });
   };
 
   delete = <T>(url: string, params?: AxiosRequestConfig['params']) => {