+++ packages/i18n/src/translations/index.ts
@@ -1,6 +1,0 @@
-import de from './de.json';
-import en from './en.json';
-import es from './es.json';
-
-const translations = { de, en, es };
-export default translations;