+++ packages/react-native-sdk/src/components/participants/ParticipantsInfoListView.tsx
@@ -7,6 +7,7 @@ import {
   Restricted,
   useCall,
   useConnectedUser,
+  useI18n,
   useParticipants,
 } from '@stream-io/video-react-bindings';
 import {
@@ -59,6 +60,7 @@ export const ParticipantsInfoListView = ({
   setIsCallParticipantsInfoViewVisible,
 }: ParticipantsInfoListViewProps) => {
   const participants = useParticipants();
+  const { t } = useI18n();
   const [selectedParticipant, setSelectedParticipant] = useState<
     StreamVideoParticipant | undefined
   >(undefined);
@@ -117,7 +119,9 @@ export const ParticipantsInfoListView = ({
             <View style={styles.header}>
               <View style={styles.leftHeaderElement} />
               <Text style={styles.headerText}>
-                Participants ({participants.length})
+                {t('Participants ({{ numberOfParticipants }})', {
+                  numberOfParticipants: participants.length,
+                })}
               </Text>
               <Pressable
                 onPress={onCloseCallParticipantsViewVisible}
@@ -130,14 +134,14 @@ export const ParticipantsInfoListView = ({
             <FlatList data={participants} renderItem={renderItem} />
             <View style={styles.buttonGroup}>
               <Pressable style={styles.button} onPress={inviteHandler}>
-                <Text style={styles.buttonText}>Invite</Text>
+                <Text style={styles.buttonText}>{t('Invite')}</Text>
               </Pressable>
               <Restricted requiredGrants={[OwnCapability.MUTE_USERS]}>
                 <Pressable
                   style={styles.button}
                   onPress={muteAllParticipantsHandler}
                 >
-                  <Text style={styles.buttonText}>Mute All</Text>
+                  <Text style={styles.buttonText}>{t('Mute All')}</Text>
                 </Pressable>
               </Restricted>
             </View>
@@ -197,7 +201,7 @@ const ParticipantInfoItem = (props: ParticipantInfoType) => {
   return (
     <Pressable style={styles.participant} onPress={optionsOpenHandler}>
       <View style={styles.participantInfo}>
-        <Avatar radius={theme.avatar.xs} participant={participant} />
+        <Avatar size={theme.avatar.xs} participant={participant} />
 
         <Text style={styles.name} numberOfLines={1}>
           {(participant.name || generateParticipantTitle(participant.userId)) +